Consolidate parsers into one file
