+++ src/utils/parser.ts
@@ -2,6 +2,7 @@ import * as babelParser from "@babel/parser";
 import { codeFrameColumns } from "@babel/code-frame";
 import { Expression, Node, ObjectExpression } from "@babel/types";
 import { getErrorLocation } from "./getErrorLocation";
+import { ChildNode, Declaration, parse, Root } from "postcss";
 
 export function parseJsObject(input: string): [ObjectExpression, string] {
   let rawLines: string;
@@ -49,4 +50,42 @@ export function nodeToString(node: Node, code: string): string {
 
   // Copy the string value directly from the input code
   return code.slice(start, end).trim();
+}
+
+/**
+ * Finds all css delcarations
+ * @param root PostCss AST Root
+ */
+function getCssProps(root: Root): Declaration[] {
+  const { nodes } = root;
+
+  if (!nodes) {
+    return [];
+  }
+
+  function iterate(nodes: ChildNode[], declarations: Declaration[] = []) {
+    nodes.forEach((item) => {
+      if (item.type === "decl") {
+        declarations.push(item);
+      } else if (item.type === "rule" && item?.nodes) {
+        iterate(item.nodes, declarations);
+      }
+    });
+
+    return declarations;
+  }
+
+  return iterate(nodes);
+}
+/**
+ * Parses css into PostCss AST and returns array of declarations
+ * @param css css string
+ */
+export function parseCss(css: string): Declaration[] {
+  try {
+    const root: Root = parse(css);
+    return getCssProps(root);
+  } catch (e) {
+    throw new Error("Could not parse CSS");
+  }
 }
